requirements.txt
opencv-python
tqdm
matplotlib
dataclasses
pyyaml
yacs
fvcore==0.1.2.post20201213
scikit-image==0.17.2
scipy==1.5.4
smplx==0.1.26
pyro-ppl==1.7.0
